changed the routing for logged users and just casual users

+++ routes/book_routes.py
@@ -1,7 +1,8 @@
 from flask import Blueprint, request, render_template, redirect, url_for, flash
 from flask_login import login_required, current_user
-from models import Book, Review, db
+from models import Book, Review, User, db
 from datetime import datetime
+from sqlalchemy.orm import joinedload
 
 book_bp = Blueprint('book', __name__)
 
@@ -11,7 +12,28 @@ def books():
     search = request.args.get('search', '')
     author_filter = request.args.get('author', '')
 
-    query = Book.query
+    query = Book.query.options(joinedload(Book.submitter))
+
+    if search:
+        query = query.filter(Book.title.contains(search))
+    if author_filter:
+        query = query.filter(Book.author.contains(author_filter))
+
+    books = query.paginate(page=page, per_page=10)
+    return render_template("books.html", books=books, search=search, author_filter=author_filter)
+
+@book_bp.route("/<username>/books")
+@login_required
+def user_books(username):
+    if current_user.username != username:
+        flash('You do not have permission to view this page.')
+        return redirect(url_for('book.user_books', username=current_user.username))
+    
+    page = request.args.get('page', 1, type=int)
+    search = request.args.get('search', '')
+    author_filter = request.args.get('author', '')
+
+    query = Book.query.options(joinedload(Book.submitter))
 
     if search:
         query = query.filter(Book.title.contains(search))
@@ -23,13 +45,28 @@ def books():
 
 @book_bp.route("/book/<int:book_id>")
 def book_detail(book_id):
-    book = Book.query.get_or_404(book_id)
-    reviews = Review.query.filter_by(book_id=book_id).all()
+    book = Book.query.options(joinedload(Book.submitter)).get_or_404(book_id)
+    reviews = Review.query.filter_by(book_id=book_id).options(joinedload(Review.author)).all()
     return render_template("book_detail.html", book=book, reviews=reviews)
 
-@book_bp.route("/add_book", methods=['GET', 'POST'])
+@book_bp.route("/<username>/book/<int:book_id>")
 @login_required
-def add_book():
+def user_book_detail(username, book_id):
+    if current_user.username != username:
+        flash('You do not have permission to view this page.')
+        return redirect(url_for('book.user_book_detail', username=current_user.username, book_id=book_id))
+    
+    book = Book.query.options(joinedload(Book.submitter)).get_or_404(book_id)
+    reviews = Review.query.filter_by(book_id=book_id).options(joinedload(Review.author)).all()
+    return render_template("book_detail.html", book=book, reviews=reviews)
+
+@book_bp.route("/<username>/add_book", methods=['GET', 'POST'])
+@login_required
+def add_book(username):
+    if current_user.username != username:
+        flash('You do not have permission to add books.')
+        return redirect(url_for('book.user_books', username=current_user.username))
+    
     if request.method == 'POST':
         title = request.form.get('title')
         author = request.form.get('author')
@@ -53,17 +90,21 @@ def add_book():
         db.session.commit()
 
         flash('Book added successfully!')
-        return redirect(url_for('book.books'))
+        return redirect(url_for('book.user_books', username=username))
 
     return render_template("add_book.html")
 
-@book_bp.route("/edit_book/<int:book_id>", methods=['GET', 'POST'])
+@book_bp.route("/<username>/edit_book/<int:book_id>", methods=['GET', 'POST'])
 @login_required
-def edit_book(book_id):
+def edit_book(username, book_id):
+    if current_user.username != username:
+        flash('You do not have permission to edit books.')
+        return redirect(url_for('book.user_books', username=current_user.username))
+    
     book = Book.query.get_or_404(book_id)
     if book.submitted_by != current_user.id:
         flash('You do not have permission to edit this book.')
-        return redirect(url_for('book.book_detail', book_id=book.id))
+        return redirect(url_for('book.user_book_detail', username=username, book_id=book.id))
     
     if request.method == 'POST':
         book.title = request.form.get('title')
@@ -76,19 +117,23 @@ def edit_book(book_id):
 
         db.session.commit()
         flash('Book updated successfully!')
-        return redirect(url_for('book.book_detail', book_id=book.id))
+        return redirect(url_for('book.user_book_detail', username=username, book_id=book.id))
 
     return render_template("edit_book.html", book=book)
 
-@book_bp.route("/delete_book/<int:book_id>", methods=['POST'])
+@book_bp.route("/<username>/delete_book/<int:book_id>", methods=['POST'])
 @login_required
-def delete_book(book_id):
+def delete_book(username, book_id):
+    if current_user.username != username:
+        flash('You do not have permission to delete books.')
+        return redirect(url_for('book.user_books', username=current_user.username))
+    
     book = Book.query.get_or_404(book_id)
     if book.submitted_by != current_user.id:
         flash('You do not have permission to delete this book.')
-        return redirect(url_for('book.book_detail', book_id=book.id))
+        return redirect(url_for('book.user_book_detail', username=username, book_id=book.id))
     
     db.session.delete(book)
     db.session.commit()
     flash('Book deleted successfully!')
-    return redirect(url_for('book.books'))
+    return redirect(url_for('book.user_books', username=username))
